Button restructure
